requirements.txt
pytest
mock
contextvars
boto3
kafka-python